Move uikit to top level, fix font loading/styles for non-vendored font files.
Museo Sans and Cubano aren't free and are vendored in prod. For development, this isn't easy to set up, but the fonts are available on Creative Cloud and elsewhere, and can be installed separately.

Using these external installs generally requires a different naming convention, which has been added to the font stack in ThemeProvider.